Do not force session to load
By forcing the session to load, a new session will be created if
one does not already exist. This means that the response sent to
the client will contain the `Set-Cookie` header with a newly
created `_session_id` value. While this behaviour is okay for
clients making their very first request and as such do not already
have an existing session, forcing the session to load results in
a subtle bug in other scenarios.

For example, it is possible for the client to already be logged in
to the application and as such, have an existing `_session_id`
cookie. However, if a request is made with a missing or invalid
Cross-Site Request Forgery (CSRF) token, the existing session will
not be available for the duration of that request. As a result, if
the session is forcefully loaded, a new session will be created
instead. This new session will overwrite the existing session due
to the `Set-Cookie` header mentioned above as the client will
update its `_session_id` cookie value to the new value returned in
the response. Because subsequent requests from the client will use
the new value of the `_session_id` cookie which corresponds to a
session that has no logged in user, the client is essentially
logged out.

The implication of the above is that a malicious actor can
forcefully log out users on all ESTL applications that make use of
this library along with cookie-based authentication.