Add missing lock file and Ruby version for reproducible builds (#21)
* Set Ruby version to v3.3.1

* Remove `Gemfile.lock` from `.gitignore`

The lock file should be checked into version control so that we
get reproducible builds. In particular, the lock file is necessary
for locking the versions of (direct & indirect) dependencies.

* Generate lock file

* Fix `test_helper` load error

cannot load such file --
/superlogger/test/superlogger_test.rb/test_helper (LoadError)

* Fix Rails version to v6

In v7, the asset pipeline gem `sprockets` was removed, leading to
the following error when running the dummy application:
undefined method `assets' for an instance of
Rails::Application::Configuration (NoMethodError)

* Fix sqlite3 version to v1.4

The latest version v2.0.2 does not work with the rest of the old
dependencies:
Error loading the 'sqlite3' Active Record adapter. Missing a gem
it depends on? can't activate sqlite3 (~> 1.4), already activated
sqlite3-2.0.2-arm64-darwin. Make sure all dependencies are added
to Gemfile. (LoadError)

* Revert changes to `test_helper`

For some reason, when using an old version of Rails, the load
error seen before no longer occurs even without the fix.

* Add instructions on running tests

* Move Rails version constraint to `Gemfile`

The version constraint is a developer dependency for the library
itself, and shouldn't act as a constraint for consumers of the
library.